refactor: relocate header file
