+++ src/components/common/Layout.tsx
@@ -2,8 +2,8 @@ import styled from '@emotion/styled';
 
 import Footer from './Footer';
 
-import Header from '@/components/common/Header';
 import SideNav from '@/components/common/SideNav';
+import Header from '@/components/header/Header';
 
 interface LayoutProps {
   children: React.ReactNode;
